Commit poner los metodos hechos en game en un scripth
-Scripts puestos en comodins
