+++ Assets/Scenes/Feaure005/JuegoComodins.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4ebfc739dd1bac547ac583756f5f7618
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
