+++ Assets/Scripts/Comodins.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 15c7494c2748d75409ef43c3adb5007a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
