+++ UserSettings/EditorUserSettings.asset
@@ -8,6 +8,9 @@ EditorUserSettings:
     RecentlyUsedScenePath-0:
       value: 22424703114646680e0b0227036c75051310176439262f2434
       flags: 0
+    RecentlyUsedScenePath-1:
+      value: 22424703114646680e0b0227036c791517020a2f7c78737f073c1834edc33b35ede33fe7f4742a323016f6
+      flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
       flags: 0
